Begin creating menus for requesting permissions
