+++ www/css/atmos-ui.css
@@ -342,6 +342,21 @@ select{
 	}
 }
 
+#android-permission-setup{
+	z-index: 99;
+	position: absolute;
+	background-color: hsla(0,0%,99%,1.00);
+	width: 100%;
+	height: 100%;
+	top: 0;
+	left: 0;
+	padding: 0;
+}
+
+#android-permission-container {
+	padding: 15px;
+}
+
 /* Dark Mode Colors */
 body.dark {
 	background-color: #222222;
@@ -366,6 +381,10 @@ body.dark {
 		left: 0;
 	}
 
+	#android-permission-setup {
+		background-color: #222222;
+	}
+
 	#welcome-window-native{
 		z-index: 100;
 		position: absolute;
@@ -376,20 +395,7 @@ body.dark {
 		left: 0;
 	}
 
-	#android-permission-setup{
-		z-index: 99;
-		position: absolute;
-		background-color: #222222;
-		width: 100%;
-		height: 100%;
-		top: 0;
-		left: 0;
-		padding: 0;
-	}
 
-	#android-permission-container {
-		padding: 15px;
-	}
 
 	.android-setup {
 		width: calc(100vw - 30px);
